Move program_jtag into commands

+++ include/urjtag/cmd.h
@@ -82,7 +82,7 @@ int urj_cmd_params (char * const params[]);
  */
 int urj_cmd_get_number (const char *s, long unsigned *i);
 
-int program_jtag(int fd, const char *drivername, const char *bsdl_path, int frequency);
+int program_jtag(int32_t fd, const char *drivername, const char *bsdl_path, int64_t frequency);
